requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.1
alembic==1.15.1
altair==4.2.2
amqp==5.3.1
annotated-types==0.7.0
ansicolors==1.1.8
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==22.2.0
auth0-python==4.8.1
azure-common==1.1.28
azure-core==1.32.0
azure-search-documents==11.5.2
Babel==2.14.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
billiard==4.2.1
bleach==5.0.1
blinker==1.9.0
blis==1.0.1
blobfile==3.0.0
build==1.2.2.post1
cachelib==0.13.0
cachetools==5.2.1
catalogue==2.0.10
celery==5.3.4
certifi==2022.12.7
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.0.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.20.0
cohere==5.14.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.0.7
crewai==0.108.0
crewai-tools==0.38.1
cryptography==44.0.2
cycler==0.11.0
cymem==2.0.8
Cython==0.29.33
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distlib==0.3.6
distro==1.9.0
Django==4.0.10
docker==7.1.0
docstring_parser==0.16
docx2txt==0.9
durationpy==0.9
embedchain==0.1.128
entrypoints==0.4
et_xmlfile==2.0.0
executing==2.0.1
fairscale==0.4.13
faiss-cpu==1.9.0
fastapi==0.115.12
fastavro==1.10.0
fastjsonschema==2.16.2
ffmpeg==1.4
ffmpy==0.4.0
filelock==3.12.0
fire==0.7.0
Flask==3.1.1
Flask-Caching==2.3.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2025.3.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.73.0
google-auth==2.38.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.69.2
gptcache==0.1.44
gradio==5.5.0
gradio_client==1.4.2
greenlet==3.0.0
groq==0.13.1
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
h5py==3.8.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.21.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
hyperframe==6.1.0
idna==2.8
importlib_metadata==8.6.1
importlib_resources==6.5.2
instructor==1.7.7
ipykernel==6.29.3
ipython==8.22.2
ipython-genutils==0.2.0
ipywidgets==8.1.2
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.8
jedi==0.19.1
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
json_repair==0.40.0
jsonpatch==1.33
jsonpickle==4.0.2
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.1.3
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab-pygments==0.2.2
jupyterlab_rise==0.42.0
jupyterlab_server==2.25.4
jupyterlab_widgets==3.0.10
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kombu==5.5.4
kubernetes==32.0.1
lancedb==0.21.2
langchain==0.3.25
langchain-cohere==0.3.5
langchain-core==0.3.64
langchain-experimental==0.3.4
langchain-openai==0.3.21
langchain-text-splitters==0.3.8
langcodes==3.4.1
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
language_data==1.2.0
libclang==16.0.0
litellm==1.60.2
lxml==5.3.0
Mako==1.3.9
marisa-trie==1.2.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.26.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mem0ai==0.1.77
mistune==2.0.4
ml-dtypes==0.1.0
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy_extensions==1.1.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.5.6
networkx==3.4.2
nodeenv==1.9.1
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.25.2
oauthlib==3.2.2
onnx==1.17.0
onnxruntime==1.21.0
onnxruntime-gpu==1.21.0
onnxruntime-tools==1.7.0
openai==1.68.2
opencv-python==4.5.5.62
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
opt-einsum==3.3.0
orjson==3.10.16
ormsgpack==1.10.0
overrides==7.7.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
papermill==2.6.0
parso==0.8.3
pdfminer.six==20231228
pdfplumber==0.11.5
pickleshare==0.7.5
pillow==10.4.0
pipenv==2023.3.20
platformdirs==2.6.2
playwright==1.39.0
portalocker==2.10.1
portpicker==1.2.0
posthog==3.23.0
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.43
propcache==0.3.1
protobuf==5.29.4
psutil==5.9.8
psycopg2-binary==2.9.10
pure-eval==0.2.2
py3nvml==0.2.7
pyarrow==19.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.8.1b0
pydub==0.25.1
pyee==11.0.1
Pygments==2.14.0
PyJWT==2.10.1
Pympler==1.0.1
PyMySQL==1.1.1
pyparsing==3.0.9
pypdf==5.4.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyright==1.1.398
pyrsistent==0.19.3
pysbd==0.3.4
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.12
python-version==0.0.2
pytube==15.0.0
pytz==2024.2
pytz-deprecation-shim==0.1.0.post0
pyvis==0.3.2
pywin32==305
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.0.0
qdrant-client==1.13.3
qtconsole==5.5.1
QtPy==2.4.1
redis==5.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.24.0
rsa==4.9
ruff==0.7.3
safehttpx==0.1.1
safetensors==0.4.5
sagemath==1.3.0
schedule==1.2.2
schema==0.7.7
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==3.3.0
sentencepiece==0.2.0
shellingham==1.5.4
simplegeneric==0.8.1
six==1.16.0
smart-open==7.0.5
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.3.2.post1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.21
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.3
starlette==0.46.1
streamlit==1.37.1
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.18.1
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.2.1
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
tomlkit==0.12.0
toolz==0.12.0
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tornado==6.3
tqdm==4.66.4
traitlets==5.14.2
transformers==4.46.2
typer==0.15.2
types-python-dateutil==2.9.0.20240316
types-requests==2.32.0.20250306
typing-inspect==0.9.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==4.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
uv==0.6.10
uvicorn==0.34.0
validators==0.20.0
vine==5.1.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
wasabi==1.1.3
watchdog==3.0.0
watchfiles==1.0.4
wcwidth==0.2.6
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.7.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.10
wrapt==1.14.1
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
youtube-transcript-api==0.6.2
zipp==3.21.0
zstandard==0.23.0
